pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "bioml"
version = "0.1.0"
description = "Binary classification toolkit for biomedical datasets"
authors = [{ name = "Scott A. Bowler", email = "sbowler49601@gmail.com" }]
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
keywords = ["machine learning", "bioinformatics", "classification", "biomarker"]

dependencies = [
    "pandas>=2.0",
    "numpy>=1.24",
    "scikit-learn>=1.3",
    "xgboost>=2.0",
    "matplotlib>=3.7",
    "seaborn>=0.12",
    "scipy>=1.11",
    "pyyaml>=6.0",
    "openpyxl>=3.1",
]

[project.optional-dependencies]
nn = ["tensorflow>=2.13"]
dev = ["pytest>=7.4", "pytest-cov>=4.1"]

[tool.setuptools.packages.find]
where = ["."]
include = ["bioml*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
